Refactor Enums into a package
