+++ src/main/java/enums/Command.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.enums;
 
 public enum Command {
     NAME,
