+++ src/main/java/enums/Result.java
@@ -1,4 +1,4 @@
-package main.java;
+package main.java.enums;
 
 public enum Result {
     SUCCESS,
