+++ src/main/java/task/TaskList.java
@@ -1,6 +1,6 @@
 package main.java.task;
 
-import main.java.*;
+import main.java.enums.Command;
 import main.java.exception.HamletException;
 import main.java.parser.Parser;
 import main.java.ui.Ui;
